fix: u64 -> u32 for janus id

+++ jarust_plugins/src/common.rs
@@ -12,5 +12,5 @@ pub enum JanusId {
     /// String Identifier
     String(String),
     /// Unsigned Integer Identifier
-    Uint(u32),
+    Uint(u64),
 }
